# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
  '/api/films':
    post:
      tags:
        - films
      security:
        - cookieAuth: []
      summary: Create a new film
      description: A new film is created by the authenticated user.
      operationId: createFilm
      requestBody:
        description: Representation of the film to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
            examples:
              Example 1:
                value:
                  title: "Matrix"
                  private: true
                  watchDate: "2013-05-01"
                  rating: 8
                  favorite: true
              Example 2:
                value:
                  title: "Inception"
                  private: false
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Film'
              examples:
                Example 1:
                  value:
                    id: 11,
                    title: "Matrix"
                    owner: 1
                    private: true
                    watchDate: "2013-05-01"
                    rating: 8
                    favorite: true
                    self: "/api/films/private/11"
                Example 2:
                  value:
                    id: 12,
                    title: "Inception"
                    owner: 2
                    private: false
                    self: "/api/films/private/12"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                empty body:
                  value:
                    errors:
                      - param: Server
                        msg: Empty request body.
                missing private:
                  value:
                    errors: 
                      - param: Server
                        msg: Private property is required. 
                missing title:
                  value:
                    errors:
                      - param: Server
                        msg: Title property is required.
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
  '/api/films/private': 
    get:
      tags:
        - films/private
      security:
        - cookieAuth: []
      summary: Retrieve the private films of the logged-in user 
      description: >-
        The private films of the logged-in user are retrieved. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPrivateFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              examples:
                Example:
                  value:
                    totalPages: 1
                    currentPage: 1
                    totalItems: 2
                    films:
                      - id: 4
                        title: "Weathering with You"
                        owner: 1
                        private: true
                        self: "/api/films/private/4"
                      - id: 11
                        title: "Matrix"
                        owner: 1
                        private: true
                        watchDate: "2013-05-01"
                        rating: 8
                        favorite: true
                        self: "/api/films/private/11"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string 
  '/api/films/private/{filmId}':
    get:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Retrieve a private film
      description: The private film with ID filmId is retrieved. This operation can be performed on the film if the user who performs the operation is the film's owner.
      operationId: getSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
              examples:
                Example:
                  value:
                    id: 4
                    title: "Weathering with You"
                    owner: 1
                    private: true
                    self: "/api/films/private/4"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                invalid parameters:
                  value:
                    errors:
                      - param: Server
                        msg: Invalid filmId, it must be a number.
                missing parameters:
                  value:
                    errors:
                      - param: Server
                        msg: filmId parameter missing
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The user is not the owner of the film.
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The film does not exist.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
    put:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Update a private film
      description: The private film with ID filmId is updated. This operation does not allow changing its visibility. This operation can be performed only by the owner.
      operationId: updateSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
            examples:
              Example 1:
                value: 
                  title: "The Godfather"
                  watchDate: "2018-05-01"
                  rating: 10
                  favorite: false
              Example 2:
                value: 
                  watchDate: "2018-05-01"
                  rating: 10
                  favorite: true
              Example 4:
                value: 
                  favorite: true
              Example 5:
                value: 
                  rating: 10
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                invalid parameters:
                  value:
                    errors:
                      - param: Server
                        msg: Invalid filmId, it must be a number.
                missing parameters:
                  value:
                    errors:
                      - param: Server
                        msg: filmId parameter missing.
                missing body: 
                  value: 
                    errors:
                      - param: Server
                        msg: empty body request, nothing to update.
                invalid body 1:
                  value:
                    errors:
                      - param: Server
                        msg: title must be a string.
                invalid body 2:
                  value:
                    errors:
                      - param: Server
                        msg: watchDate must be a string in the format yyyy-mm-dd.
                invalid body 3:
                  value:
                    errors:
                      - param: Server
                        msg: rating must be an integer between 1 and 10
                invalid body 4:
                  value:
                    errors:
                      - param: Server
                        msg: favorite must be a boolean value
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The user is not the owner of the film.
        '409':
          description: Conflict (attempt to make unallowed change, e.g., visibility)
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The visibility of the film cannot be changed.
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The film does not exist.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
    delete:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Delete a private film
      description: The private film with ID filmId is deleted. This operation can only be performed by the owner.
      operationId: deleteSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                invalid parameters:
                  value:
                    errors:
                      - param: Server
                        msg: Invalid filmId, it must be a number.
                missing parameters:
                  value:
                    errors:
                      - param: Server
                        msg: filmId parameter missing
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The user is not the owner of the film.
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The film does not exist.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
  '/api/films/public': 
    get:
      tags:
        - films/public
      summary: Retrieve the public films
      description: >-
        The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPublicFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              examples:
                Example:
                  value:
                    totalPages: 3
                    currentPage: 1
                    totalItems: 6
                    films:
                      - id: 1
                        title: "Your Name"
                        owner: 1
                        private: false
                        watchDate: "2021-10-03"
                        rating: 9
                        favorite: true
                        self: "/api/films/public/1"
                      - id: 2
                        title: "Heaven's Feel"
                        owner: 2
                        private: false
                        self: "/api/films/public/2"
                    next": "/api/films/public?pageNo=2"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
  '/api/films/public/invited': 
    get:
      tags:
        - films/public/invited
      security:
        - cookieAuth: []
      summary: Retrieve the public films that the logged-in user has been invited to review
      description: >-
        The public films that the logged-in user has been invited to review are retrieved. A pagination mechanism is implemented to limit the size of messages.
      operationId: getInvitedFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              examples:
                Example:
                  value: 
                    totalPages: 2
                    currentPage: 1
                    totalItems: 3
                    films:
                      - id: 3
                        title: "You Can (Not) Redo"
                        owner: 1
                        private: false
                        self: "/api/films/public/3"
                      - id: 6
                        title: "Spirited Away"
                        owner: 5
                        private: false
                        self: "/api/films/public/6"
                    next: "/api/films/public/invited?pageNo=2"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
...
  '/api/users':
    get:
      tags:
        - users
      security:
        - cookieAuth: []
      summary: Get information about the users
      description: >-
        The available information (passwords excluded) about all the
        users is retrieved. This operation is available only to authenticated users.
      operationId: getUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              examples:
                Example:
                  value:
                    - id: 1
                      name: "User"
                      email: "user.dsp@polito.it"
                      self: "/api/users/1"
                    - id: 2
                      name: "Frank"
                      email: "frank.stein@polito.it"
                      self: "/api/users/2"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '404':
          description: Not Found
          content: {}
...
components:
  schemas:
...
    Film:
      description: schema for Film data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description: true if only the owner of the film data structure can access it, false otherwise
          default: 'true'
        watchDate:
          type: string
          format: date
          description: the date when the film was watched by the owner
        rating:
          type: integer
          description: the rating the owner has given to the film
          minimum: 1
          maximum: 10
        favorite:
          type: boolean
          description: 'true if the film is among the favourite
            ones of the owner, false otherwise '
          default: 'false'
        self: 
          type: string
          format: uri
      required:
      - title
      - private
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid